+++ src/Pages/handbooks/InfoPage.tsx
@@ -33,17 +33,24 @@ const InfoPage: React.FC = () => {
   };
 
   const handleInfoDelete = async (id: string) => {
-    toast.promise(
-      deleteInfo(id),
-      {
-        pending: 'Processing your request...',
-        success: 'Info has been deleted Successfully 👌',
-        error: 'Error encountered 🤯',
-      },
-      {
-        theme: 'dark',
+    Modal.confirm({
+      title: "Подтверждаете удаление записи?",
+      okText: "Да",
+      cancelText: "Нет",
+      okType: "danger",
+      onOk: () => {
+        toast.promise(
+          deleteInfo(id),
+          {
+            pending: 'Processing your request...',
+            success: 'Info has been deleted Successfully 👌',
+            error: 'Error encountered 🤯',
+          },
+          {
+            theme: 'dark',
+          });
       }
-    );
+    })
   };
 
   const showModal = (item: InfoModel | null) => {
@@ -107,7 +114,7 @@ const InfoPage: React.FC = () => {
       key: 'imageId',
       render: (imageId: string) =>
         <Image src=
-          {imageId ? `http://localhost:7152/api/storage/${imageId}` : require("../../Assets/nocontent.png")}
+          {`http://localhost:7152/api/storage/${imageId}`}
           alt="no content"
           style={{ width: '100%', maxWidth: '120px' }} />,
     },
@@ -134,12 +141,12 @@ const InfoPage: React.FC = () => {
           <Row>
             <Col xs={24} md={{ span: 16, offset: 4 }}>
               <div className="d-flex align-items-center justify-content-between mb-4">
-                <h1 className="text-success">Новости</h1>
+                <h1 style={{ fontFamily: "cursive" }}>Новости</h1>
                 <Button type="primary" icon={<PlusOutlined />} onClick={() => showModal(null)}>
-                  Add Info
+                  Добавить
                 </Button>
               </div>
-              <Table dataSource={data} columns={columns} rowKey="id" />
+              <Table className="custom-table" dataSource={data} columns={columns} rowKey="id" />
             </Col>
           </Row>
           <Modal
@@ -148,18 +155,18 @@ const InfoPage: React.FC = () => {
             onCancel={handleCancel}
             footer={null}
           >
-            <Form<InfoModel> form={form} onFinish={onFinish} layout="vertical" >
+            <Form<InfoModel> form={form} onFinish={onFinish} layout="vertical">
               <Form.Item name="title" label="Заголовк" rules={[{ required: true }]}>
-                <Input maxLength={50}/>
+                <Input maxLength={50} />
               </Form.Item>
               <Form.Item name="text" label="Текст" rules={[{ required: true }]}>
-                <TextArea rows={8} maxLength={5000}/>
+                <TextArea rows={8} maxLength={5000} />
               </Form.Item>
               <Form.Item name="date" label="Дата" rules={[{ required: true }]}>
                 <Input type='date' />
               </Form.Item>
               <Form.Item name="imageId" label="Ссылка на картинку" rules={[{ required: true }]}>
-                <Input maxLength={50}/>
+                <Input maxLength={50} />
               </Form.Item>
               <ImageUploader onImageIdChange={handleImageIdChange} />
               <Form.Item>
